Move tag rendering to render::tags
