+++ src/render.rs
@@ -1,22 +1,19 @@
 pub mod filters;
+pub mod tags;
 pub mod types;
 
 use std::borrow::Cow;
 use std::collections::HashMap;
 
-use pyo3::exceptions::PyAttributeError;
 use pyo3::prelude::*;
-use pyo3::types::{PyDict, PyList};
 
 use crate::error::{PyRenderError, RenderError};
-use crate::parse::{Tag, TagElement, TokenTree, Url};
-use crate::template::django_rusty_templates::NoReverseMatch;
+use crate::parse::{TagElement, TokenTree};
 use crate::types::Argument;
 use crate::types::ArgumentType;
 use crate::types::TemplateString;
 use crate::types::Text;
 use crate::types::Variable;
-use crate::utils::PyResultMethods;
 use types::{Content, Context};
 
 pub type ResolveResult<'t, 'py> = Result<Option<Content<'t, 'py>>, PyRenderError>;
@@ -104,100 +101,6 @@ impl Resolve for Variable {
             object_at.1 += key_at.1 + 1;
         }
         Ok(Some(Content::Py(variable)))
-    }
-}
-
-fn current_app(py: Python, request: &Option<Py<PyAny>>) -> PyResult<Py<PyAny>> {
-    let none = py.None();
-    let request = match request {
-        None => return Ok(none),
-        Some(request) => request,
-    };
-    if let Ok(current_app) = request
-        .getattr(py, "current_app")
-        .ok_or_isinstance_of::<PyAttributeError>(py)?
-    {
-        return Ok(current_app);
-    }
-    match request
-        .getattr(py, "resolver_match")
-        .ok_or_isinstance_of::<PyAttributeError>(py)?
-    {
-        Ok(resolver_match) => resolver_match.getattr(py, "namespace"),
-        Err(_) => Ok(none),
-    }
-}
-
-impl Resolve for Url {
-    fn resolve<'t, 'py>(
-        &self,
-        py: Python<'py>,
-        template: TemplateString<'t>,
-        context: &mut Context,
-    ) -> ResolveResult<'t, 'py> {
-        let view_name = match self.view_name.resolve(py, template, context)? {
-            Some(view_name) => view_name,
-            None => Content::String(Cow::Borrowed("")),
-        };
-        let urls = py.import("django.urls")?;
-        let reverse = urls.getattr("reverse")?;
-
-        let current_app = current_app(py, &context.request)?;
-        let url = if self.kwargs.is_empty() {
-            let py_args = PyList::empty(py);
-            for arg in &self.args {
-                py_args.append(arg.resolve(py, template, context)?)?;
-            }
-            reverse.call1((
-                view_name,
-                py.None(),
-                py_args.to_tuple(),
-                py.None(),
-                current_app,
-            ))
-        } else {
-            let kwargs = PyDict::new(py);
-            for (key, value) in &self.kwargs {
-                kwargs.set_item(key, value.resolve(py, template, context)?)?;
-            }
-            reverse.call1((view_name, py.None(), py.None(), kwargs, current_app))
-        };
-        match &self.variable {
-            None => Ok(Some(Content::Py(url?))),
-            Some(variable) => match url.ok_or_isinstance_of::<NoReverseMatch>(py)? {
-                Ok(url) => {
-                    context.context.insert(variable.clone(), url.unbind());
-                    Ok(None)
-                }
-                Err(_) => Ok(None),
-            },
-        }
-    }
-}
-
-impl Render for Tag {
-    fn render<'t>(
-        &self,
-        py: Python<'_>,
-        template: TemplateString<'t>,
-        context: &mut Context,
-    ) -> RenderResult<'t> {
-        Ok(match self {
-            Self::Autoescape { enabled, nodes } => {
-                let autoescape = context.autoescape;
-                context.autoescape = enabled.into();
-
-                let mut rendered = vec![];
-                for node in nodes {
-                    rendered.push(node.render(py, template, context)?)
-                }
-
-                context.autoescape = autoescape;
-                Cow::Owned(rendered.join(""))
-            }
-            Self::Load => Cow::Borrowed(""),
-            Self::Url(url) => url.render(py, template, context)?,
-        })
     }
 }
